IGNITE-3487: applied review comments

+++ modules/indexing/src/main/java/org/apache/ignite/internal/processors/query/h2/opt/GridH2Table.java
@@ -899,7 +899,8 @@ public class GridH2Table extends TableBase {
 
                     idxs.remove(i);
 
-                    if (idx instanceof GridH2ProxyIndex)
+                    if (idx instanceof GridH2ProxyIndex &&
+                        idx.getSchema().findIndex(session, idx.getName()) != null)
                         database.removeSchemaObject(session, idx);
 
                     continue;
